rust: remove x86 panicking intrinsics.
This removes the panicking implementation of the x86 intrinsics.

This requires a PR [0] for the rust core library to remove the need for
the `__udivti3` intrinsic. This PR disables formatting support for the
`i/u128` rust primitives when the `no_iu128_fmt` config option is set.

In order to remove the need for the other intrinsics a recent rust
version is required (Tested with 1.84).

Link: https://github.com/rust-lang/rust/pull/136385 [0]
Signed-off-by: Christian Schrefl <chrisi.schrefl@gmail.com>